+++ CHANGELOG.md
@@ -13,6 +13,15 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+### Removed 
+- account_created, created and owned attributes to DictKey resource
+- account_number and branch_code to PaymentPreview resource
+### Changed
+- account_number and branch_code docstring attributes to DictKey resource
+### Fixed
+- account_type docstring attribute to DictKey resource
+### Added
+- update function to Deposit resource
 
 ## [2.6.2] - 2021-11-10
 ### Changed